requirements.txt
streamlit==1.24.0
pandas==2.0.3
plotly==5.15.0
robin-stocks==2.1.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
pytest==8.0.0
httpx>=0.24.0  # For Discord webhook integration
aiofiles>=0.7.0  # For async file operations
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3
PyJWT>=2.8.0

# Development Dependencies
pylint==2.17.4
